+++ docs/changes.md
@@ -1,5 +1,6 @@
 2022/01/24
   * 优化tk\_strncpy（感谢智明提供补丁）
+  * 完善fscript debugger
 
 2022/01/23
   * 完善fscript debugger,支持attach/launch两种启动方式。